requirements.txt
langchain==0.2.15
black==24.8.0
isort==5.13.2
python-dotenv==1.0.1
streamlit==1.38.0
langchain-community==0.2.15
langchain-openai==0.1.23
langgraph==0.2.15
streamlit-mermaid==0.2.0